Refactor : OAuth dto 그룹화
