+++ src/main/java/com/ee06/wooms/domain/users/dto/oauth/GithubResponse.java
@@ -1,24 +1,24 @@
-package com.ee06.wooms.domain.users.dto;
+package com.ee06.wooms.domain.users.dto.oauth;
 
 import com.ee06.wooms.domain.users.entity.SocialProvider;
 
 import java.util.Map;
 
-public class GoogleResponse implements OAuth2Response{
+public class GithubResponse implements OAuth2Response {
     private final Map<String, Object> attributes;
 
-    public GoogleResponse(Map<String, Object> attributes) {
+    public GithubResponse(Map<String, Object> attributes) {
         this.attributes = attributes;
     }
 
     @Override
     public String getProvider() {
-        return SocialProvider.GOOGLE.getSocialProvider();
+        return SocialProvider.GITHUB.getSocialProvider();
     }
 
     @Override
     public String getProviderId() {
-        return attributes.get("sub").toString();
+        return attributes.get("id").toString();
     }
 
     @Override
@@ -31,4 +31,3 @@ public class GoogleResponse implements OAuth2Response{
         return attributes.get("name").toString();
     }
 }
-
