+++ src/main/java/com/ee06/wooms/domain/users/dto/oauth/GoogleResponse.java
@@ -1,24 +1,24 @@
-package com.ee06.wooms.domain.users.dto;
+package com.ee06.wooms.domain.users.dto.oauth;
 
 import com.ee06.wooms.domain.users.entity.SocialProvider;
 
 import java.util.Map;
 
-public class GithubResponse implements OAuth2Response {
+public class GoogleResponse implements OAuth2Response{
     private final Map<String, Object> attributes;
 
-    public GithubResponse(Map<String, Object> attributes) {
+    public GoogleResponse(Map<String, Object> attributes) {
         this.attributes = attributes;
     }
 
     @Override
     public String getProvider() {
-        return SocialProvider.GITHUB.getSocialProvider();
+        return SocialProvider.GOOGLE.getSocialProvider();
     }
 
     @Override
     public String getProviderId() {
-        return attributes.get("id").toString();
+        return attributes.get("sub").toString();
     }
 
     @Override
@@ -31,3 +31,4 @@ public class GithubResponse implements OAuth2Response {
         return attributes.get("name").toString();
     }
 }
+
